+++ src/mkvtoolnix-gui/util/recently_used_strings.h
@@ -21,6 +21,7 @@ public:
 
   bool isEmpty() const;
 
+  int maximumNumItems() const;
   QStringList items() const;
   void setItems(QStringList const &newItems);